requirements.txt
PyQt6==6.6.1
pyserial==3.5
sounddevice==0.4.6
requests==2.31.0
python-dotenv==1.0.0
pyttsx3==2.90
numpy==1.26.4
edge-tts==6.1.9
gTTS==2.5.1
pygame==2.5.2
aiofiles==24.1.0
